+++ core/src/main/java/net/okocraft/box/core/model/data/BoxCustomDataContainer.java
@@ -3,7 +3,7 @@ package net.okocraft.box.core.model.data;
 import com.github.siroshun09.configapi.api.Configuration;
 import net.okocraft.box.api.model.data.CustomDataContainer;
 import net.okocraft.box.core.storage.model.data.CustomDataStorage;
-import net.okocraft.box.core.util.InternalExecutors;
+import net.okocraft.box.core.util.executor.InternalExecutors;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
