+++ core/src/main/java/net/okocraft/box/core/model/manager/BoxUserManager.java
@@ -3,7 +3,7 @@ package net.okocraft.box.core.model.manager;
 import net.okocraft.box.api.model.manager.UserManager;
 import net.okocraft.box.api.model.user.BoxUser;
 import net.okocraft.box.core.storage.model.user.UserStorage;
-import net.okocraft.box.core.util.InternalExecutors;
+import net.okocraft.box.core.util.executor.InternalExecutors;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Optional;
